fix(H_9235): fork_a_matrix xor_t float64 bitwise_xor 버그 — .astype(int)
code[]가 concepts.json서 float64 로드라 code[a]^code[b]가 bitwise_xor unsupported.
fork_a_precheck.py의 xor()와 동일하게 .astype(int) 캐스트 추가.

+++ state/g1_gamma_binding_lane/fork_a_matrix.py
@@ -45,7 +45,7 @@ T, D = Z[keys[0] + "__seq"].shape
 
 
 def xor_t(a, b):
-    return (code[a] ^ code[b]).astype(np.float64)
+    return (code[a].astype(int) ^ code[b].astype(int)).astype(np.float64)
 
 
 def sigmoid(x):
